+++ README.md
@@ -21,7 +21,7 @@ Contained in [db_tasks.py](./tasks/db_tasks.py)
 ## Configuration Export Tasks
 Contained in [export_tasks.py](./tasks/export_tasks.py)
 
-The configuration output is generated by the [config_generator](./app/config_generator/README.md) module. This module contains functions to generate fund and round configuration, form JSONs, and HTML representations for a given funding round.
+The configuration output is generated by the [config_generator](./app/export/README.md) module. This module contains functions to generate fund and round configuration, form JSONs, and HTML representations for a given funding round.
 
 ## Database
 ### Schema
