--- app/config_generator/output/R605/fund_store/fund_config_09-08-2024.py
@@ -1,13 +1,0 @@
-{
-    "id": "c80bd7ce-91a7-4e99-8944-01798953df1c",
-    "short_name": "SFF863",
-    "welsh_available": False,
-    "owner_organisation_name": "Department for Fishing",
-    "owner_organisation_shortname": "DF",
-    "owner_organisation_logo_uri": "http://www.google.com",
-    "name_json": {"en": "Salmon Fishing Fund"},
-    "title_json": {"en": "funding to improve access to salmon fishing"},
-    "description_json": {
-        "en": "A £10m fund to improve access to salmon fishing facilities across the devolved nations."
-    },
-}
