--- app/config_generator/output/R605/fund_store/round_config_09-08-2024.py
@@ -1,39 +1,0 @@
-{
-    "id": "debde85d-2b1d-4c88-bec5-6eeaa735e6ea",
-    "fund_id": "c80bd7ce-91a7-4e99-8944-01798953df1c",
-    "short_name": "R605",
-    "opens": "2024-08-09T09:12:05.697258",
-    "assessment_start": "2024-08-09T09:12:05.697259",
-    "deadline": "2024-08-09T09:12:05.697259",
-    "application_reminder_sent": False,
-    "reminder_date": "2024-08-09T09:12:05.697260",
-    "assessment_deadline": "2024-08-09T09:12:05.697260",
-    "prospectus": "http://www.google.com",
-    "privacy_notice": "http://www.google.com",
-    "reference_contact_page_over_email": False,
-    "contact_email": None,
-    "contact_phone": None,
-    "contact_textphone": None,
-    "support_times": None,
-    "support_days": None,
-    "instructions_json": None,
-    "feedback_link": None,
-    "project_name_field_id": None,
-    "application_guidance_json": None,
-    "guidance_url": None,
-    "all_uploaded_documents_section_available": None,
-    "application_fields_download_available": None,
-    "display_logo_on_pdf_exports": None,
-    "mark_as_complete_enabled": None,
-    "is_expression_of_interest": None,
-    "eoi_decision_schema": None,
-    "feedback_survey_config": {
-        "has_feedback_survey": None,
-        "has_section_feedback": None,
-        "is_feedback_survey_optional": None,
-        "is_section_feedback_optional": None,
-    },
-    "eligibility_config": {"has_eligibility": None},
-    "title_json": {"en": "round the first"},
-    "contact_us_banner_json": {"en": "", "cy": ""},
-}
