--- app/config_generator/output/R605/fund_store/sections_config_09-08-2024.py
@@ -1,13 +1,0 @@
-[
-    {"section_name": {"en": "1. Organisation Information", "cy": ""}, "tree_path": "12.1", "requires_feedback": None},
-    {
-        "section_name": {"en": "1.1 About your organisation", "cy": ""},
-        "tree_path": "12.1.1",
-        "form_name_json": {"en": "about-your-org", "cy": ""},
-    },
-    {
-        "section_name": {"en": "1.2 Contact Details", "cy": ""},
-        "tree_path": "12.1.2",
-        "form_name_json": {"en": "contact-details", "cy": ""},
-    },
-]
