--- app/config_generator/output/R605/html/full_application.html
@@ -1,81 +1,0 @@
-<div class="govuk-!-margin-bottom-8">
-  <h2 class="govuk-heading-m ">
-    Table of contents
-  </h2>
-  <ol class="govuk-list govuk-list--number">
-    <li>
-      <a class="govuk-link" href="#organisation-information">
-        Organisation Information
-      </a>
-    </li>
-  </ol>
-  <hr class="govuk-section-break govuk-section-break--l govuk-section-break--visible" />
-  <h2 class="govuk-heading-l" id="organisation-information">
-    1. Organisation Information
-  </h2>
-  <h3 class="govuk-heading-m">
-    1.1. About your organisation
-  </h3>
-  <h4 class="govuk-heading-s">
-    1.1.1. Organisation Name
-  </h4>
-  <p class="govuk-body">
-    What is your organisation's name?
-  </p>
-  <p class="govuk-body">
-    This must match the regsitered legal organisation name
-  </p>
-  <p class="govuk-body">
-    Does your organisation use any other names?
-  </p>
-  <p class="govuk-body">
-    If 'No', go to <strong>1.1.2</strong>
-  </p>
-  <p class="govuk-body">
-    If 'Yes', go to <strong>1.1.1.2</strong>
-  </p>
-  <h4 class="govuk-heading-s">
-    1.1.1.2. Alternative names of your organisation
-  </h4>
-  <p class="govuk-body">
-    Alternative Name 1
-  </p>
-  <h4 class="govuk-heading-s">
-    1.1.2. Organisation Address
-  </h4>
-  <p class="govuk-body">
-    What is your organisation's address?
-  </p>
-  <p class="govuk-body">
-    This must match the regsitered organisation address
-  </p>
-  <h4 class="govuk-heading-s">
-    1.1.3. Organisation Classification
-  </h4>
-  <p class="govuk-body">
-    How is your organisation classified?
-  </p>
-  <ul class="govuk-list govuk-list--bullet">
-    <li class="">
-      Charity
-    </li>
-    <li class="">
-      Public Limited Company
-    </li>
-  </ul>
-  <h3 class="govuk-heading-m">
-    1.2. Contact Details
-  </h3>
-  <h4 class="govuk-heading-s">
-    1.2.1. Lead Contact Details
-  </h4>
-  <p class="govuk-body">
-    Main Contact Name
-  </p>
-  <p class="govuk-body">
-    Main Contact Email
-  </p>
-  <p class="govuk-body">
-    Main Contact Address
-  </p>
-</div>
