+++ app/export/config_generator/generate_fund_round_form_jsons.py
@@ -2,10 +2,10 @@ import json
 
 from flask import current_app
 
-from app.config_generator.generate_form import build_form_json
-from app.config_generator.scripts.helpers import validate_json
-from app.config_generator.scripts.helpers import write_config
 from app.db.queries.round import get_round_by_id
+from app.export.config_generator.generate_form import build_form_json
+from app.export.config_generator.helpers import validate_json
+from app.export.config_generator.helpers import write_config
 
 form_schema = {
     "$schema": "http://json-schema.org/draft-07/schema#",
