+++ app/export/config_generator/generate_fund_round_html.py
@@ -1,10 +1,10 @@
 from flask import current_app
 
 from app.all_questions.metadata_utils import generate_print_data_for_sections
-from app.config_generator.generate_all_questions import print_html
-from app.config_generator.generate_form import build_form_json
-from app.config_generator.scripts.helpers import write_config
 from app.db.queries.round import get_round_by_id
+from app.export.config_generator.generate_all_questions import print_html
+from app.export.config_generator.generate_form import build_form_json
+from app.export.config_generator.helpers import write_config
 
 
 def generate_all_round_html(round_id):
