+++ app/export/config_generator/helpers.py
@@ -1,6 +1,4 @@
 import os
-from dataclasses import asdict
-from dataclasses import is_dataclass
 from datetime import date
 
 import jsonschema
@@ -9,20 +7,12 @@ from jsonschema import validate
 
 from app.blueprints.self_serve.routes import human_to_kebab_case
 from app.blueprints.self_serve.routes import human_to_snake_case
-
-
-def convert_to_dict(obj):
-    if is_dataclass(obj):
-        return asdict(obj)
-    elif isinstance(obj, list):
-        return [asdict(item) if is_dataclass(item) else item for item in obj]
-    else:
-        return obj
+from app.shared.helpers import convert_to_dict
 
 
 def write_config(config, filename, round_short_name, config_type):
     # Determine the base output directory
-    base_output_dir = f"app/config_generator/output/{round_short_name}/"
+    base_output_dir = f"app/export/config_generator/output/{round_short_name}/"
 
     if config_type == "form_json":
         output_dir = os.path.join(base_output_dir, "form_runner/")
