+++ app/shared/helpers.py
@@ -1,0 +1,11 @@
+from dataclasses import asdict
+from dataclasses import is_dataclass
+
+
+def convert_to_dict(obj):
+    if is_dataclass(obj):
+        return asdict(obj)
+    elif isinstance(obj, list):
+        return [asdict(item) if is_dataclass(item) else item for item in obj]
+    else:
+        return obj
