+++ tests/test_build_assessment_config.py
@@ -1,4 +1,4 @@
-from app.config_generator.scripts.generate_assessment_config import (
+from app.export.config_generator.generate_assessment_config import (
     build_assessment_config,
 )
 from tests.unit_test_data import cri1
@@ -8,7 +8,7 @@ from tests.unit_test_data import mock_form_1
 
 def test_build_basic_structure(mocker):
     mocker.patch(
-        "app.config_generator.scripts.generate_assessment_config.get_form_for_component", return_value=mock_form_1
+        "app.export.config_generator.generate_assessment_config.get_form_for_component", return_value=mock_form_1
     )
 
     results = build_assessment_config([cri1])
@@ -20,7 +20,7 @@ def test_build_basic_structure(mocker):
 def test_with_field_info(mocker):
 
     mocker.patch(
-        "app.config_generator.scripts.generate_assessment_config.get_form_for_component", return_value=mock_form_1
+        "app.export.config_generator.generate_assessment_config.get_form_for_component", return_value=mock_form_1
     )
     results = build_assessment_config([cri1])
     assert len(results["unscored_sections"]) == 1
