+++ tests/test_generate_form.py
@@ -2,16 +2,16 @@ from uuid import uuid4
 
 import pytest
 
-from app.config_generator.generate_form import build_conditions
-from app.config_generator.generate_form import build_form_json
-from app.config_generator.generate_form import build_lists
-from app.config_generator.generate_form import build_navigation
-from app.config_generator.generate_form import build_page
-from app.config_generator.generate_form import human_to_kebab_case
 from app.db.models import Component
 from app.db.models import ComponentType
 from app.db.models import Lizt
 from app.db.models import Page
+from app.export.config_generator.generate_form import build_conditions
+from app.export.config_generator.generate_form import build_form_json
+from app.export.config_generator.generate_form import build_lists
+from app.export.config_generator.generate_form import build_navigation
+from app.export.config_generator.generate_form import build_page
+from app.export.config_generator.generate_form import human_to_kebab_case
 from tests.unit_test_data import mock_c_1
 from tests.unit_test_data import mock_form_1
 
@@ -73,7 +73,7 @@ def test_human_to_kebab(input, exp_output):
 )
 def test_build_lists(mocker, pages, exp_result):
     mocker.patch(
-        "app.config_generator.generate_form.get_list_by_id",
+        "app.export.config_generator.generate_form.get_list_by_id",
         return_value=Lizt(
             name="greetings_list",
             type="string",
@@ -610,7 +610,7 @@ def test_build_navigation_no_conditions(mocker, input_partial_json, input_pages,
 )
 def test_build_navigation_with_conditions(mocker, input_pages, input_partial_json, exp_next, exp_conditions):
     mocker.patch(
-        "app.config_generator.generate_form.build_page",
+        "app.export.config_generator.generate_form.build_page",
         return_value={"path": "/organisation-alternative-names", "next": []},
     )
     results = build_navigation(partial_form_json=input_partial_json, input_pages=input_pages)
